// ClientApp/src/app/tools-management/models/Acceptance.ts
export interface Acceptance {
    acceptanceId: number;
    procurementId: number;
    demandId: number;
    demandTypeId:number;
    itemCategoryId:number;
    itemDetailId:number;
    departmentNameId: number;
    departmentName:string;
    itemDetail:string;
    itemName:string;
    qty:number;
    condition:string;
    remarks:string;
    sparesCategoryId:number;
    procurementStatusId: number;
    sourceOfSupplyId:number;
    manufactureId:number;
    principalNameId: number;
    placeOfDeliveryId: number;
    demandAuthorityId:number;
    conditionOfItemId: number;
    itemInspectionId:number;
    sftDate: Date;
    sftLetterNo:string;
    workOrderNo: string;
    workOrderDate: Date;
    sftQty:number;
    storeQty:number;
    itemSerNo:string;
    model:string;
    brand: string;
    warranty: string;
    warrantyFrom:Date;
    warrantyTo:Date;
    deliveryDate:Date;
    inspectionDate: Date;
    purchasePrice:string;
    dateOfManufacture: Date;
    acDocument: string;
    arcDocument:string;
    cofcDocument: string;
    othersDocument: string;
    sftRegPage:string;
    acceptanceDocument:string;
    docVerification: string;
    sftStatus: boolean;
    menuPosition:string;
    isActive: boolean;
   
}